+++ src/main/kotlin/spock/adb/actions/BaseAction.kt
@@ -1,0 +1,43 @@
+package spock.adb.actions
+
+import com.android.ddmlib.IDevice
+import com.intellij.notification.NotificationType
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.ui.popup.PopupChooserBuilder
+import com.intellij.ui.components.JBList
+import org.jetbrains.android.sdk.AndroidSdkUtils
+import spock.adb.AdbController
+import spock.adb.AdbControllerImp
+import spock.adb.notification.CommonNotifier.Companion.showNotifier
+
+abstract class BaseAction : AnAction() {
+    override fun actionPerformed(event: AnActionEvent) = event.project?.run {
+        val controller = AdbControllerImp(this, AndroidSdkUtils.getDebugBridge(this))
+        controller.connectedDevices { list ->
+            if (list.isNotEmpty())
+                showDeviceList(project = this, list) {
+                    performAction(controller, it)
+                }
+            else
+                showNotifier(project = this, content = "No Devices", type = NotificationType.ERROR)
+
+        }
+    } ?: cancelAction(event)
+
+    private fun showDeviceList(project: Project, devices: List<IDevice>, block: (device: IDevice) -> Unit) {
+        val list = JBList(devices.map { it.name })
+        PopupChooserBuilder(list).apply {
+            this.setTitle("Devices")
+            this.setItemChoosenCallback {
+                block(devices[list.selectedIndex])
+            }
+            this.createPopup().showCenteredInCurrentWindow(project)
+        }
+
+    }
+
+    private fun cancelAction(event: AnActionEvent) {}
+    abstract fun performAction(controller: AdbController, device: IDevice)
+}