+++ src/main/kotlin/spock/adb/actions/RestartAppAction.kt
@@ -1,0 +1,10 @@
+package spock.adb.actions
+
+import com.android.ddmlib.IDevice
+import spock.adb.AdbController
+
+class RestartAppAction : BaseAction() {
+    override fun performAction(controller: AdbController, device: IDevice) {
+        controller.restartApp(device)
+    }
+}