+++ src/main/kotlin/spock/adb/premission/ListItem.kt
@@ -1,0 +1,3 @@
+package spock.adb.premission
+
+data class ListItem(var permission: String, var isSelected: Boolean = false)