+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 # Spock Adb Changelog
 ## [Unreleased]
 ### Added
-
+- Get Current App BackStack (Activities and nested fragments)
+- Add Plugins actions ex GetCurrentFragment,RestartApp,etc
 ### Changed
 
 ### Deprecated
@@ -12,6 +13,9 @@
 
 ### Fixed
 
+- support latest version of AS
+- fix get current fragment
+- fix If two instances of AS are open, the plugin does not work properly
 ### Security
 ## [1.0.9]
 ### Added